kai_commit_miner/pyproject.toml
[project]
name = "kai-commit-miner"
version = "0.1.0"
description = "Mine commit history of migrated projects to extract migration knowledge"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
  "kai-mcp-solution-server",
  "httpx>=0.27.0",
  "langchain>=1.2.4",
  "langchain-core>=1.2.7",
  "langchain-openai>=1.1.7",
  "langchain-ollama>=1.0.1",
  "pydantic>=2.10.6",
  "pydantic-settings>=2.9.1",
  "pyyaml>=6.0.1",
  "langchain-anthropic>=1.4.0",
]

[dependency-groups]
dev = [
  "coverage>=7.4.0",
  "mypy>=1.8.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.0",
]

[tool.uv.sources]
kai-mcp-solution-server = { path = "../kai_mcp_solution_server" }

[build-system]
requires = ["uv_build>=0.7.2,<0.8"]
build-backend = "uv_build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py312"
